refactor : NotificationItem 분리
